feat #63 Refactors ImportScanner to get imports for a class
Create a new class for the old inner class ImportScanner,
CtImportScanner, and his new interface, CtImport, to have a
contract to retrieve all imports for a CtSimpleType element
(represent a class, an interface or an annotation). And
add a new test case to check if we are all useful imports.