Supports lambda expressions.
Adds CtLambda in meta model and builds it in JDTTreeBuilder.
A lambda can have a type, parameters (typed or not) and a body
with statements or with an expression. Finally, adds tests case
to test multi syntaxes of lambda.
Adds signature too. When we have a reference of a CtLambda and
want a signature of this element, we print something like
(Predicate<T>) (p). We display the type of the lambda and their
parameters.