fix bug: NPE when compilation units array contains null elements
Fix a bug in TreeBuilderCompiler.buildUnits() where a retured array is
longer then expected and the extra spaces are null. The method is updated
to use the _totalUnits_ instance variable as the true element count in
the _unitsToProcess_ array instead of the length of the array itself.

This broke processing of classes using JDTBatchCompiler.getUnits() and
expecting every element in the array to be non-null.  The problem came up
when trying to process the source for a large GWT project.