Enhanced access lookup in lambda for noclasspath mode.
Until now, a token within a lambda (noclasspath mode) is always
considered a variable access (in particular a CtLocalVariable), though,
it also could refer to a type---a class for instance. This commit adds
two features:

1) An improved variable lookup method which ensures that a found
variable is visible in current context. Furthermore, the amount of
cloned code has been reduced.

2) Check if a given token (SingleNameReference) is a variable or type
access and create an appropriate AST node instead of always creating
a local variable reference.