fixes bug in printing of new Array with comment #911
In case of comment the last character is the last character of the
comment and not the " , ". So we have to call printer.removeLastChar();
before comment is added. 
It is not enough to skip adding " , ", because printer.removeLastChar();
removes EOL added by previous call of scan() too. It influences
formatting of spooned code