change from empty types to a flag for removed.
If a type is deleted(there parent == null), a CtType is flagged as removed. If a CtType gets a parent the CtType removed flag is set to false.
This allows skipping all deleted files and not only empty files. We define the flag by parent, because a Type without parent is not part of the AST.