add strict mode.
We can need to check some cases after the model is build. Otherwise the api gets way harder to use.
e.g. localVariableRef.setSimpleName(List<String>[]) is in strict mode only a legal identifier, if the type reference refers to
an arry (isArray= true). In non strict mode a user can set the name and afterwards set the isArray to true