chore: Try to list new errors in javadoc regression check
Before this patch the CI Javadoc regression check would only look at the
amount of violations and fail the build if they increased:

```
JAVADOC QUALITY SCORE (lower is better)
    Compare: 2584
    Current: 2592

Javadoc quality has deteriorated!
Run the chore/ci-checkstyle-javadoc.sh script locally to find errors
See https://github.com/inria/spoon/issues/3923 for details
```

This wasn't really helpful when actually finding the errors, as you
would need to run the checkstyle command locally on both branches and
diff the output yourself. This patch adds a small python script that
tries to (somewhat intelligently) perform this diff in the CI and print
a more useful failure message.

The line numbers in the output will not be correct if the exact same
violation message appears multiple times in a single file, but it is
better than having no feedback.