Don't clone method body when adapting it for isOverriding
Cloning is quite slow and method bodies can be large - this results in
the isOverriding check taking quite a bit longer than needed.

Removing the type adaption is not possible as the javadoc explains: If
we do not adapt the method, we have no CtTypeParameter declaration
lying around our references can attach to. This results in
getTypeErasure() potentially entering an infinite recursion.