fix: Use a better heuristic for guessing type/field references in no classpath mode
In no classpath mode, JDT only gives us a ProblemBinding with a
QualifiedNameReference. We then have to figure out whether this is a
type, field, inner class, package or anything else ourselves.
Previously, every child of an invocation was treated as a type. This
patch changes it to favor fields and only falls back to a type access,
if it appeared somewhere in an import.

This is still far from perfect, but it at least produces better guesses.

Closes #5151