Refactor code to fix runtime exception caused by package name with CtType.INNERTTYPE_SEPARATOR.
The check for the CtType.INNERTTYPE_SEPARATOR that caused runtime exceptions has been removed so that valid packages with CtType.INNERTTYPE_SEPARATOR can be found.

A new test has been added to verify that a package with a name containing a dollar sign can be successfully retrieved using the PackageFactory.