release: Skip GPG plugin and remove js git actions
This commit includes the addition of a token in JReleaser's Github checkout. It also adds a run command that creates a new branch, commits the release version changes, and pushes the branch. Furthermore, `-Dgpg.skip=true` is added to the run command for skipping gpg verification. Lastly, the `Commit & Push changes` step is modified by removing the using action and adding commands for committing the changes and merging the branch with master.