Add @Nullable annotation to TypeBinding parameters
Added @Nullable annotation to two functions in ReferenceBuilder.java and JDTTreeBuilderHelper.java. These functions are getVariableReference() in ReferenceBuilder and declaringType variable in JDTTreeBuilderHelper. The Nullable annotation is useful as the TypeBinding parameter can at times be null, particularly when running the program in 'no classpath' mode. This reduces the risk of throwing a null pointer exception and increases overall code robustness.