Disable AssignmentToNull check in qodana.yaml
Disabled the AssignmentToNull check in qodana.yaml due to its lack of support for nullability annotations. Commented out the line to maintain code clarity for future reference. This change enhances program performance when running in 'no classpath' mode and prevents potential NullPointerException. It contributes to the overall robustness of the code.