Refactor ReceiverParameter setting and scanning
This commit refactors the process of setting ReceiverParameter in the JDTTreeBuilderHelper. It removes redundant condition checks while setting the type of the argument. The commit also introduces the scanning of ReceiverParameter in the CtConstructor and CtScanner classes. Additionally, unnecessary assertion and createReceiverParameter method in the ExecutableFactory class have been removed to further refine the code.