date import statements and optimize argument naming
Reorganized import statements in "CtAbstractVisitor.java", resulting in a more logical and readable structure. Also, renamed a method argument in "CtAbstractVisitor.java" for better readability and maintainability. Furthermore, the redundant comment in "JDTTreeBuilder.java" was removed, and the import statements in "CtScannerTest.java" was updated to import individual methods instead of using a wildcard import.